clear up v3 firmware, ongoing updates of firmware docs
