+++ coresensors/v2/docs/sensorStreamFormat_v2.tex
@@ -2,6 +2,8 @@
 
 \usepackage{longtable}
 
+\usepackage{indentfirst}
+
 \usepackage[utf8]{inputenc}
 \usepackage[T1]{fontenc}
 \usepackage{fancyhdr}
@@ -34,7 +36,7 @@
         keywordstyle=\color{keywords},
         commentstyle=\color{comments},
         stringstyle=\color{red},
-        showstringspaces=false,
+        % showstringspaces=false,
         identifierstyle=\color{green},
         procnamekeys={def,class}}
 
@@ -45,31 +47,33 @@
 
 \geometry{top = 1in, bottom = 1in, left = 0.5in, right = 0.5in}
 
-\setlength\parindent{0pt}
+\setlength{\parindent}{4em}
+\setlength{\parskip}{1em}
+\renewcommand{\baselinestretch}{1.5}
 
 \pagestyle{fancy}
 \renewcommand{\footrulewidth}{1pt}
 \lhead{Waggle Sensor Array}
 \chead{Interface and Data Format Specification}
-\rhead{Version 0.4 alpha 1}
+\rhead{Version 1.01}
 \lfoot{Waggle Group}
 \rfoot{park708@purdue.edu}
 
 \setlength{\tabcolsep}{10pt}
-\renewcommand{\arraystretch}{1.4}
+\renewcommand{\arraystretch}{1.0}
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
 \begin{document}
 
 \begin{titlepage}
-   \vspace*{\stretch{1.0}}
-   \begin{center}
+  \vspace*{\stretch{1.0}}
+  \begin{center}
         \Huge\textbf{\textsc{Interface and \\ Data Format Specification \\ for sensors \\
         \Large\textsc{(v2 sensor boards)}}}\\[0.5cm]
         \Large\textsc{Waggle Group \\ Waggle Sensor Array}\\[1cm]
-        \large\textsc{July 2016, }
-        \large\textsc{Version 0.4 alpha 1}\\
+        \large\textsc{November 2016, }
+        \large\textsc{Version 1.01}\\
    \end{center}
    \vspace*{\stretch{2.5}}
 \end{titlepage}
