+++ coresensors/v3/RainDoc/PacketExample.tex
@@ -1,0 +1,121 @@
+\section{Example}
+
+In the example below, the spaces and brackets are used to clearly show the different fields within the packet.  The actual packet, or any component of it, does not have spaces or brackets.  This example is in hex format.
+
+%%%%%%%%%%%%%%%%%%%%
+
+\subsection{Whole Packet}
+
+\{AA\} \hspace{0.25em} \{0\} \hspace{0.25em} \{C2\} \hspace{0.25em} \{[0 86 6 5 4 3 2 1] \hspace{0.25em} [1 82 F1 49] \hspace{0.25em} [2 4 C8 2C CE 8] \hspace{0.25em} [3 2 55 8C] \hspace{0.25em} [4 5 B3 83 D1 98 54] \hspace{0.25em} [5 82 2D 5] \hspace{0.25em} [6 2 75 6B] \hspace{0.25em} [7 8 88 A3 92 8C E2 A1 8D CF] \hspace{0.25em} [8 82 30 D1] \hspace{0.25em} [9 2 6D 61] \hspace{0.25em} [A 6 CC C2 84 99 82 9C] \hspace{0.25em} [B 4 B4 9E AA C2] \hspace{0.25em} [C 2 69 C5] \hspace{0.25em} [D 82 30 C4] \hspace{0.25em} [E 82 A E5] \hspace{0.25em} [F 82 4D 39] \hspace{0.25em} [10 2 77 E8] \hspace{0.25em} [11 A2 85 F ED 81 F3 5 DB 62 C3 5 A0 42 E1 26 ED E1 B4 25 99 61 D0 88 81 CB BA AA C0 C D5 5E AB C4 B2 6] \hspace{0.25em} [12 2 B5 52] \hspace{0.25em} [13 82 DD 5F] \hspace{0.25em} [14 82 3D CE] \hspace{0.25em} [15 82 72 61] \hspace{0.25em} [16 2 E 78] \hspace{0.25em} [17 82 17 E] \hspace{0.25em} [18 2 62 7] \hspace{0.25em} [19 82 5E 47] \hspace{0.25em} [1A 82 58 29] \hspace{0.25em} [1B 2 5B A2] \hspace{0.25em} [1C 82 62 46] \hspace{0.25em} [1D 4 9C 2C C7 1A] \hspace{0.25em} [1E 3 B3 12 64] \hspace{0.25em} [1F 86 C B A 9 8 7] \hspace{0.25em} [FE 84 FF DF FF FF]\} \hspace{0.25em} \{29\} \hspace{0.25em} \{55\}\\
+
+To avoid any confusion, this is what the packet actually looks like:\\
+
+\seqsplit{
+AA0C2086654321182F14924C82CCE832558C45B383D198545822D562756B7888A3928CE2A18DCF88230D1926D61A6CCC28499829CB4B49EAAC2C269C5D8230C4E82AE5F824D3910277E811A285FED81F35DB62C35A042E126EDE1B4259961D08881CBBAAAC0CD55EABC4B26122B5521382DD5F14823DCE15827261162E78178217E18262719825E471A8258291B25BA21C8262461D49C2CC71A1E3B312641F86CBA987FE84FFDFFFFF2955}
+
+\begin{table}[H]
+    \centering
+    {\rowcolors{2}{black!12}{black!2}
+    \begin{tabular}{|c c c|}
+        \hline
+        \textbf{Field} & \textbf{Value} & \textbf{Explanation}\\
+        \hline
+        \hline
+        Start & 0xAA & Start-of-packet indicator\\
+        Protocol version & 0x00 & Version of Waggle sensor protocol being used\\
+        Length of data & 0xC2 & Byte length of data sub-packet\\
+        Data & & Data sub-packet\\
+        CRC & 0x29 & CRC-8 of data sub-packet\\
+        End & 0x55 & End-of-packet indicator\\
+        \hline
+    \end{tabular}
+    }
+    \caption{Example of whole packet}
+\end{table}
+
+%%%%%%%%%%%%%%%%%%%%
+
+\subsection{Data Sub-Packet}
+
+\begin{table}[H]
+    \centering
+    {\rowcolors{2}{black!12}{black!2}
+    \begin{tabular}{|c c|}
+        \hline
+        \textbf{Value} & \textbf{Explanation}\\
+        \hline
+        \hline
+        0x00 & ID\\
+        0x86 & 0b\textbf{\textit{1}}0000000 (valid) OR'd with 0x06 (length)\\
+        0x06 & Address byte 0\\
+        0x05 & Address byte 1\\
+        0x04 & Address byte 2\\
+        0x03 & Address byte 3\\
+        0x02 & Address byte 4\\
+        0x01 & Address byte 5\\
+        \hline
+    \end{tabular}
+    }
+    \caption{Example of MAC address}
+\end{table}
+
+\begin{table}[H]
+    \centering
+    {\rowcolors{2}{black!12}{black!2}
+    \begin{tabular}{|c c|}
+        \hline
+        \textbf{Value} & \textbf{Explanation}\\
+        \hline
+        \hline
+        0x01 & ID\\
+        0x82 & 0b\textbf{\textit{1}}0000000 (valid) OR'd with 0x02 (length)\\
+        0xF1 & 0b\textbf{\textit{1}}0000000 (data converted) OR'd with 0x71 (integer)\\
+        0x49 & 0b\textbf{\textit{0}}0000000 (positive) OR'd with 0x49 (fractional)\\
+        \textit{113.73} & Temperature in decimal\\
+        \hline
+    \end{tabular}
+    }
+    \caption{Example of TMP112}
+\end{table}
+
+\begin{table}[H]
+    \centering
+    {\rowcolors{2}{black!12}{black!2}
+    \begin{tabular}{|c c|}
+        \hline
+        \textbf{Value} & \textbf{Explanation}\\
+        \hline
+        \hline
+        0x1E & ID\\
+        0x03 & 0b\textbf{\textit{0}}0000000 (not valid) OR'd with 0x03 (length)\\
+        0xB3 & 0b\textbf{\textit{1}}0000000 (data converted) OR'd with 0b0\textbf{\textit{0}}000000 (positive) OR'd with 0b00\textbf{\textit{110011}} (6 MSb of data)\\
+        0x12 & Middle byte of data\\
+        0x64 & LSB of data\\
+        \textit{5118100} & Pressure in decimal (pascals)\\
+        \hline
+    \end{tabular}
+    }
+    \caption{Example of Bosh}
+\end{table}
+
+\begin{table}[H]
+    \centering
+    {\rowcolors{2}{black!12}{black!2}
+    \begin{tabular}{|c c|}
+        \hline
+        \textbf{Value} & \textbf{Explanation}\\
+        \hline
+        \hline
+        0x0FE & ID\\
+        0x84 & 0b\textbf{\textit{1}}0000000 (valid) OR'd with 0x04 (length)\\
+        0xFF & Sensors (including MAC addresses) 31-24 are healthy (operating correctly)\\
+        0xDF & Sensor 21 (total reducing gases) not healthy. Sensors 23, 22, 20-16 are healthy\\
+        0xFF & Sensors 15-8 are healthy\\
+        0xFF & Sensors 7-0 are healthy\\
+        \hline
+    \end{tabular}
+    }
+    \caption{Example of sensor status (health)}
+\end{table}
+
+%%%%%%%%%%%%%%%%%%%%
