+++ coresensors/v3/RainDoc/RainDoc.tex
@@ -1,0 +1,87 @@
+\documentclass{article}
+
+\usepackage{longtable}
+
+\usepackage{indentfirst}
+
+\usepackage[utf8]{inputenc}
+\usepackage[T1]{fontenc}
+\usepackage{fancyhdr}
+\usepackage{geometry}
+\usepackage{array}
+\usepackage[usenames,dvipsnames,svgnames,table]{xcolor}
+\usepackage{multirow}
+\usepackage{morefloats}
+\usepackage{float}
+\usepackage{soul}
+\usepackage{color}
+\usepackage{seqsplit}
+\usepackage{gensymb}
+\usepackage{siunitx}
+\usepackage{graphicx}
+\usepackage{pdfpages}
+\usepackage[framemethod=TikZ]{mdframed}
+\setcounter{secnumdepth}{3}
+
+\usepackage[procnames]{listings}
+\usepackage{color}
+
+\definecolor{keywords}{RGB}{255,0,90}
+\definecolor{comments}{RGB}{0,0,113}
+\definecolor{red}{RGB}{160,0,0}
+\definecolor{green}{RGB}{0,150,0}
+
+\lstset{language=Python,
+        basicstyle=\ttfamily\small,
+        keywordstyle=\color{keywords},
+        commentstyle=\color{comments},
+        stringstyle=\color{red},
+        % showstringspaces=false,
+        identifierstyle=\color{green},
+        procnamekeys={def,class}}
+
+
+\restylefloat{table}
+
+\fontfamily{cmss}\selectfont
+
+\geometry{top = 1in, bottom = 1in, left = 0.5in, right = 0.5in}
+
+\setlength{\parindent}{4em}
+\setlength{\parskip}{1em}
+\renewcommand{\baselinestretch}{1.5}
+
+\pagestyle{fancy}
+\renewcommand{\footrulewidth}{1pt}
+\lhead{Waggle Sensor Array}
+\chead{Rain Gauge}
+\rhead{Version 0.1 alpha 1}
+\lfoot{Waggle Group}
+\rfoot{park708@purdue.edu}
+
+\setlength{\tabcolsep}{10pt}
+\renewcommand{\arraystretch}{1.0}
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+\begin{document}
+
+\begin{titlepage}
+   \vspace*{\stretch{1.0}}
+   \begin{center}
+        \Huge\textbf{\textsc{Interface and \\ Data Format Specification \\ for Rain Gauge \\
+        \Large\textsc(v3 sensor boards and rain gauge)}}\\[0.5cm]
+        \Large\textsc{Waggle Group \\ Waggle Sensor Array}\\[1cm]
+        \large\textsc{November 2016, }
+        \large\textsc{Version 0.1 alpha 1}\\
+   \end{center}
+   \vspace*{\stretch{2.5}}
+\end{titlepage}
+
+\tableofcontents
+\newpage
+
+\input{interfaceImg.tex}
+\input{packetStructure.tex}
+\input{sensorDataUnits.tex}
+\end{document}
