+++ coresensors/v3/RainDoc/packetStructure.tex
@@ -1,0 +1,68 @@
+\newpage
+
+\section{Data Transmission} \label{section:overall}
+
+The data from the sensor boards is sent as a formatted unit of data -- a transmission
+packet. A transmission packet is composed of several data sub-packets, each
+of which carries information pertaining to the parameter listed in the sub-packet.
+The data sub-packet, especially for rain gauge, is described here.
+
+\subsection{Rain gauge Data Sub-packets} \label{ssec:sub-pack}
+
+The data segment of the transmission packet is further broken down into many
+sub-packets. The sub-packet starts with a source identifier. One bit
+validity field and seven bits ``length of the sub-packet'' field
+are packed together as the next byte. The length field counts the number of
+bytes following it which make up the sub-packet. The table below shows the organization
+of a rain gauge sub-packet. If the field is set to 1, it indicates a valid measurement/reading, 
+and if the validity bit is set to 0 if the sensor represented in the sub-packet
+is dead, disabled, unconnected, unresponsive or if data could not be collected
+from the sensor in the time window. In this case, the particular invalid sub-packet is not packed into a transmission packet.
+\par
+As shown in Table \ref{table:rain_subpacket}, sensor identifier of rain gauge is 0xFC, which is 252 in decimal,
+and if the data are valid, second byte will be 0x84, which means the sub-packet is valid and length of the sub-packet is
+4 Bytes including sensor idetifier and valitity Byte.
+
+
+\begin{table}[H]
+    \centering {
+    \begin{tabular}{|c|c|c|c|}
+        \hline
+        \rowcolor{black!8}
+        \textbf{Source ID} & \bf{1-bit Validity [0: invalid, 1: valid] | 7-bit Data Length} & \textbf{Data} \\
+        \rowcolor{black!8}
+        (1 Byte) & (1 Byte) & (2 Bytes) \\
+        \hline
+        0xFC & 0x84 (valid) & count of pendant event\\ 
+        \hline
+    \end{tabular}
+    }
+    \caption{Sub-packet for rain gauge}
+    \label{table:rain_subpacket}
+\end{table}
+
+
+\subsection{Data Formats}
+
+The data sent in each sub-packet is encoded in one or more formats. Currently we define eight formats for various types of data including integers, bytes,
+and floating point numbers. Rain gauge data encoded as format 1, which is integers in range 0 to 65535, as shown in Table \ref{table:rain_format}.
+Therefore the data of the rain gauge is an integer, and it means the count of pendant event inside the rain gauge. One event of the pendant means 0.01 in. (0.254 mm) of precipitation.
+\\
+
+
+\begin{table}[H]
+    \centering {
+    \begin{tabular}{|c|c|c|c|}
+        \hline
+        \rowcolor{black!8}
+        \textbf{Format} & \textbf{Number of Bytes Used} & \textbf{Value Represented} & \textbf{Value Range} \\
+        \hline
+        1 & 2 & unsigned int\_16 input & 0 -- 65535 \\ %& MSByte LSByte\\
+        \hline
+    \end{tabular}
+    }
+    \caption{Data format for rain gauge}
+    \label{table:rain_format}
+\end{table}
+
+\clearpage
