+++ coresensors/v3/RainDoc/sensorDataUnits.tex
@@ -1,0 +1,27 @@
+\section{Sensor Data Units}\label{section:parameterUnits}
+
+The outputs of rain gauge need to be processed to meaningful values.
+The Table \ref{table:rain_data} shows the unit for the sensor values.
+`Raw Units' in the table means the unit for the packtized data, which you can get directly from the coresense boards, 
+and `Processed Units' means the unit which can be used after data conversion.
+The conversion method for the rain gauge is provided as comments in the table.
+\\
+
+
+\begin{table}[H]
+    \centering {
+    \begin{tabular}{|c|c|c|c|c|}
+        \hline
+        \rowcolor{black!8}
+        \textbf{Sensor} & \textbf{Sensor ID} & \textbf{Raw Units} & \textbf{Processed Units} & \textbf{Comments} \\ 
+        \hline
+        \multirow{2}{*}{Rain Gauge} & \multirow{2}{*}{0xFC} & \multirow{2}{*}{integer} & \multirow{2}{*}{in. or mm} & One pendant event = 0.01 in. = 0.254 mm \\
+        & & & & precipitation = output $\times$ 0.01 (or 0.254) \\ 
+        \hline
+    \end{tabular}
+    }
+    \caption{Sub-packet for rain gauge}
+    \label{table:rain_data}
+\end{table}
+
+
