+++ coresensors/v3/integrated/firmware/airsense.ino
@@ -155,7 +155,7 @@ void airsense_acquire (void)
         if (Temp_float[0] > 255 || Temp_float[0] < -63) 
         {
             TMP112[1] = (0 << 7) | LENGTH_FORMAT6;
-            SerialUSB.println("TMP112: I2C not available");
+            // SerialUSB.println("TMP112: I2C not available");
         }
         else
             TMP112[1] = (1 << 7) | LENGTH_FORMAT6;
@@ -180,7 +180,10 @@ void airsense_acquire (void)
         if ((Temp_float[1] == 998 && HTU21D_float[0] == 998) || HTU21D_float[1] < 0 || HTU21D_float[1] > 100)
         {
             HTU21D_array[1] = (0 << 7) | (LENGTH_FORMAT6 * 2);
+
+            #ifdef SERIAL_DEBUG
             SerialUSB.println("HTU21D: I2C not available");
+            #endif
         }
         else
             HTU21D_array[1] = (1 << 7) | (LENGTH_FORMAT6 * 2);
@@ -217,7 +220,10 @@ void airsense_acquire (void)
             if (Temp_long == 16777215 || BMP180_long > 110000)   // pressure, if pressure is 16777215, then I2C is diconnected
             {
                 BMP180[1] = (0 << 7) | (LENGTH_FORMAT6 + LENGTH_FORMAT5);
+
+                #ifdef SERIAL_DEBUG
                 SerialUSB.println("BMP180: I2C not available");
+                #endif
             }
             else
                 BMP180[1] = (1 << 7) | (LENGTH_FORMAT6 + LENGTH_FORMAT5);
@@ -260,7 +266,10 @@ void airsense_acquire (void)
         if (Temp_float[3] <= 0 || MMA_float[3] < 0.9)
         {
             MMA8452Q[1] = (0 << 7) | (LENGTH_FORMAT6 * 4);
+
+            #ifdef SERIAL_DEBUG
             SerialUSB.println("MMA8452Q: I2C not available");
+            #endif
         }
         else
             MMA8452Q[1] = (1 << 7) | (LENGTH_FORMAT6 * 4);
@@ -313,7 +322,10 @@ void airsense_acquire (void)
         if (Temp_float[0] > 213)
         {
             TSYS01[1] = (0 << 7) | LENGTH_FORMAT6;
+
+            #ifdef SERIAL_DEBUG
             SerialUSB.println("TSYS01: I2C not available");
+            #endif
         }
         else
             TSYS01[1] = (1 << 7) | LENGTH_FORMAT6;
