+++ coresensors/v3/integrated/firmware/buildinfo.cpp
@@ -1,2 +1,2 @@
-#define BUILD_TIME 1479488366
-#define BUILD_GIT "e9ab"
+#define BUILD_TIME 1481128101
+#define BUILD_GIT "94b2"
