+++ coresensors/v3/integrated/firmware/lightsense.ino
@@ -73,7 +73,10 @@ void lightsense_acquire (void)
     if (event.magnetic.z > 127 || event.magnetic.z > 30)
     {
         HMC5883L[1] = (0 << 7) | (LENGTH_FORMAT8 * 3);
+
+        #ifdef SERIAL_DEBUG
         SerialUSB.println("HMC5883L: I2C not available");
+        #endif
     }
     else
         HMC5883L[1] = (1 << 7) | (LENGTH_FORMAT8 * 3);
@@ -117,7 +120,10 @@ void lightsense_acquire (void)
     if (Temp_float[0] > 255 || Temp_float[1] < 0)
     {
         HIH6130[1] = (0 << 7) | (LENGTH_FORMAT6 * 2);
+
+        #ifdef SERIAL_DEBUG
         SerialUSB.println("HIH6130: I2C not available");
+        #endif
     }
     else
         HIH6130[1] = (1 << 7) | (LENGTH_FORMAT6 * 2);
@@ -152,7 +158,10 @@ void lightsense_acquire (void)
     if (Temp_uint16 == 65535)
     {
         APDS9006020[1] = (0 << 7) | LENGTH_FORMAT1;
+
+        #ifdef SERIAL_DEBUG
         SerialUSB.println("APDS9006020: I2C not available");
+        #endif
     }
     else
         APDS9006020[1] = (1 << 7) | LENGTH_FORMAT1;
@@ -179,7 +188,10 @@ void lightsense_acquire (void)
     if (Temp_uint16 == 65535)
     {
         TSL260RD[1] = (0 << 7) | LENGTH_FORMAT1;
+
+        #ifdef SERIAL_DEBUG
         SerialUSB.println("TSL260RD: I2C not available");
+        #endif
     }
     else
         TSL260RD[1] = (1 << 7) | LENGTH_FORMAT1;
@@ -205,7 +217,10 @@ void lightsense_acquire (void)
     if (Temp_uint16 == 65535)
     {
         TSL250RD_2[1] = (0 << 7) | LENGTH_FORMAT1;
+
+        #ifdef SERIAL_DEBUG
         SerialUSB.println("TSL250RD_2: I2C not available");
+        #endif
     }
     else
         TSL250RD_2[1] = (1 << 7) | LENGTH_FORMAT1;
@@ -231,7 +246,10 @@ void lightsense_acquire (void)
     if (Temp_uint16 == 65535)
     {
         MLX75305[1] = (0 << 7) | LENGTH_FORMAT1;
+
+        #ifdef SERIAL_DEBUG        
         SerialUSB.println("MLX75305: I2C not available");
+        #endif
     }
     else
         MLX75305[1] = (1 << 7) | LENGTH_FORMAT1;
@@ -256,7 +274,10 @@ void lightsense_acquire (void)
     if (Temp_uint16 == 65535)
     {
         ML8511[1] = (0 << 7) | LENGTH_FORMAT1;
+
+        #ifdef SERIAL_DEBUG
         SerialUSB.println("ML8511: I2C not available");
+        #endif
     }
     else
         ML8511[1] = (1 << 7) | LENGTH_FORMAT1;
@@ -280,7 +301,10 @@ void lightsense_acquire (void)
     if (abs(Temp_float[0]) > 127)
     {
         TMP421[1] = (0 << 7) | LENGTH_FORMAT6;
+
+        #ifdef SERIAL_DEBUG
         SerialUSB.println("TMP421: I2C not available");
+        #endif
     }
     else
         TMP421[1] = (1 << 7) | LENGTH_FORMAT6;
