+++ coresensors/v3/testing/avg/header/buildinfo.cpp
@@ -1,2 +1,2 @@
-#define BUILD_TIME 1479498494
-#define BUILD_GIT "e9ab"
+#define BUILD_TIME 1480976672
+#define BUILD_GIT "94b2"
