--- coresensors/v3/testing/software/coresense.py
@@ -1,15 +1,0 @@
-import time
-import threading
-import commands
-from coresense_InotifyKernel import *
-
-if __name__ == "__main__":
-    commL = communicatorLoader()
-    Monitor('/dev')
-    print "--------------------------------------"
-    commL.stop()
-    while len(threading.enumerate()) > 1:
-        print threading.enumerate()
-        time.sleep(1)
-        print "shutting down..."
-    commands.getoutput("rm *.pyc")
